fix the upgrade steps
